requirements.txt
django == 2.2.6
djangorestframework == 3.10.3
